+++ apps/rahat-ui/src/sections/projects/aa-2/ivr/components/ivr.dashboard.tsx
@@ -134,7 +134,7 @@ export default function IvrDashboard() {
             </p>
           </div>
         </div>
-        <CreateIVRDialog onIVRCreated={() => {}} />
+        <CreateIVRDialog />
       </div>
 
       <div className="flex flex-col sm:flex-row gap-4 mb-4">
