+++ apps/rahat-ui/src/sections/projects/aa-2/ivr/types/ivr.flow.types.ts
@@ -42,38 +42,3 @@ export interface IvrListItem {
   itemCount: number;
   lastModified: number;
 }
-
-export function findNodeById(root: IvrFlowNode, id: string): IvrFlowNode | null {
-  if (root.id === id) return root;
-  for (const child of root.children) {
-    const found = findNodeById(child, id);
-    if (found) return found;
-  }
-  return null;
-}
-
-export function flattenOptions(node: IvrFlowNode): { digit: string; label: string }[] {
-  return node.children.map((child) => ({
-    digit: child.digit || '?',
-    label: child.label,
-  }));
-}
-
-export const DIAL_PAD = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '*', '0', '#'];
-
-export function buildApiPayload(flow: IvrFlow): IvrFlowApiPayload {
-  const mapNode = (node: IvrFlowNode): IvrFlowOption => ({
-    digit: parseInt(node.digit || '0') || 0,
-    destination: node.destination || '',
-    prompt: node.prompt || '',
-    hangup: node.hangup || false,
-    options: (node.children || []).map(mapNode),
-  });
-
-  return {
-    main: {
-      prompt: flow.rootMenu.prompt || '',
-      options: (flow.rootMenu.children || []).map(mapNode),
-    },
-  };
-}
